Move the libanki.Utils.convertStreamToString() method to HtmlGenerator
Only used in HtmlGenerator, didn't really belong to libanki.
Seems generic enough to maybe move it to FileUtil?
